Replace `WalletExt` with `TxBuilderExt`
Directly building the PSBT makes it impossible to access the `TxBuilder`
that is used to construct the 2-out-of-2 multisig output. In order to
do advanced configuration like setting blocked UTXOs, we need access to
the `TxBuilder`.

Given that `maia` is a low-level protocol library, we can achieve this
by exposing more detail about how `PartyParams` is constructed and putting
some more of that "burden" onto the caller.

Most notably, our dummy multi-sig script still remains private to the
protocol as we rely on this script to correctly construct the lock transaction.